查询显示项目的模块数，用例数

+++ src/main/java/org/test/tdc/dao/rowmapper/FunctionRowMapper.java
@@ -9,8 +9,8 @@ import org.test.tdc.pojo.FunctionTO;
 public class FunctionRowMapper implements RowMapper<Object>  {
 	public Object mapRow(ResultSet rs, int rowNum) throws SQLException {
 		FunctionTO function = new FunctionTO();
-		function.setId(rs.getLong("N_ID"));
-		function.setProjectId(rs.getLong("N_PROJECT_ID"));
+		function.setId(rs.getInt("N_ID"));
+		function.setProjectId(rs.getInt("N_PROJECT_ID"));
 		function.setName(rs.getString("S_NAME"));
 		function.setCreate(rs.getDate("D_CREATE"));
